requirements.txt
aiogram>=3.0.0
easyocr>=1.7.0
gspread>=5.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
aiohttp>=3.8.0
typing-extensions>=4.0.0
